refactor: 비슷한 기능끼리 패키지로 묶기
