+++ src/main/kotlin/view/manualpick/ManualPickInput.kt
@@ -1,3 +1,5 @@
-package view
+package view.manualpick
+
+import view.LottoNumberParsedResult
 
 class ManualPickInput(val lottoNumbers: List<LottoNumberParsedResult>)
